+++ R/online.R
@@ -213,7 +213,7 @@ CheckGWS <- function(x, model, age, verbose=TRUE){
 	# limit to model
 	gwsMod <- gws[which(gws$model==model), ]
 
-	if(nrow(gwsMod)==0) stop("The selected model is not a registered output of theof the GPlates Web Service.")
+	if(nrow(gwsMod)==0) stop("The selected model is not a registered output of the GPlates Web Service.")
 
 	# limit to feature
 	feat <- gwsMod[which(gwsMod$feature==x), ]
@@ -302,28 +302,17 @@ checkgws <- function(silent=FALSE){
 # @param model The model name string
 # @param domain The domain argument, structure of velocity field
 # @param type format of the velocities
-# @param A data.frame containng the velocity information
-gwsVelocities <- function(age, model="MERDITH2021", domain="longLatGrid", type="MagAzim", verbose=FALSE){
+# @param check logical flag indicating whether the model should be checked to gws object
+# @return A data.frame containng the velocity information
+gwsVelocitiesThis <- function(x, age, model="MERDITH2021", domain="longLatGrid", type="MagAzim", verbose=FALSE, check=TRUE){
 
 	if(! requireNamespace("geojsonsf", quietly=TRUE)) stop("This method requires the 'geojsonsf' package to run.")
 
-	# we need this to check what kind of plates are there
-	data(gws, envir=environment(), package="rgplates")
+	# check for the right combination
+	if(check) CheckGWS(x=x,model=model, age=age, verbose=verbose)
 
-	# what plates are available?
-	currentModel <- gws[which(gws$model==model), ]
+	this <- paste0("velocity/",x)
 
-	# if there are plate_polygons, that has to be used, otherwise it has to static_polygons
-#	topological <- as.logical(sum(grepl("plate_polygons", currentModel$feature)))
-
-	#check domain, type and output
-
-	# depending on the type
-#	if (topological){
-#		this <- "velocity/plate_polygons"
-#	}else{
-		this <- "velocity/static_polygons"
-#	}
 	#download and save data
 	url <- paste0(gwsURL, this, '/')
 	query <- sprintf('?time=%f&model=%s&velocity_type=%s&domain_type=%s', age, model, type, domain)