+++ tests/online_velocities/test_online_velocities_plates.R
@@ -3,12 +3,18 @@ library(tinytest)
 # The current model
 currModel <- "MERDITH2021"
 
+# defaulting to static polygons
+expect_error(velocities("plate_polygons", age=30, model=currModel, type="MagAzim", domain="longLatGrid"))
+expect_error(velocities("garbage", age=30, model=currModel, type="MagAzim", domain="longLatGrid"))
+expect_silent(llgFirst <- velocities(age=30, model=currModel, type="MagAzim", domain="longLatGrid"))
+
 ################################################################################
 # I. Longlat-domain
 
 ## MagAzim
 ## data.frame
-expect_silent(llg <- velocities(age=30, model=currModel, type="MagAzim", domain="longLatGrid"))
+expect_silent(llg <- velocities("static_polygons", age=30, model=currModel, type="MagAzim", domain="longLatGrid"))
+expect_equal(llgFirst, llg) # should be the same as before
 expect_true(inherits(llg, "data.frame"))
 expect_equal(ncol(llg),5)
 expect_equal(colnames(llg),c("long", "lat", "magnitude", "azimuth", "plateid"))
@@ -25,34 +31,34 @@ expect_true(max(llg$magnitude)<=200)
 try(detach("package:terra"), silent=TRUE)
 
 # default
-expect_silent(llgrastDef <- velocities(age=30, model=currModel, type="MagAzim",
+expect_silent(llgrastDef <- velocities("static_polygons", age=30, model=currModel, type="MagAzim",
 	domain="longLatGrid", output="SpatRaster"))
 expect_true(inherits(llgrastDef, "SpatRaster"))
 
-library(terra)
+suppressPackageStartupMessages(library(terra))
 
-### resample=TRUE (default)
-expect_silent(llgrast <- velocities(age=30, model=currModel, type="MagAzim",
-	domain="longLatGrid", output="SpatRaster", cellraster=TRUE))
+### polecrop=TRUE (default)
+expect_silent(llgrast <- velocities("static_polygons", age=30, model=currModel, type="MagAzim",
+	domain="longLatGrid", output="SpatRaster", polecrop=TRUE))
 expect_equal(llgrastDef, llgrast)
-expect_equal(dim(llgrast)[3], 2)
-expect_equal(names(llgrast), c("magnitude", "azimuth"))
-expect_equal(terra::ext(llgrast), terra::ext(c(-180, 180, -90,90)))
+expect_equal(dim(llgrast)[3], 3)
+expect_equal(names(llgrast), c("magnitude", "azimuth", "plateid"))
+expect_equal(terra::ext(llgrast), terra::ext(c(-180.5, 180.5, -89.5,89.5)))
 expect_true(min(values(llgrast[["azimuth"]]), na.rm=TRUE)>=0)
 expect_true(max(values(llgrast[["azimuth"]]), na.rm=TRUE)<=2*pi)
 expect_true(min(values(llgrast[["magnitude"]]), na.rm=TRUE)>=0)
 expect_true(max(values(llgrast[["magnitude"]]), na.rm=TRUE)<=200)
 
-### resample=FALSE
-expect_silent(llgrastOrig <- velocities(age=30, model=currModel, type="MagAzim",
-	domain="longLatGrid", output="SpatRaster", cellraster=FALSE))
+### polecrop=FALSE
+expect_silent(llgrastOrig <- velocities("static_polygons", age=30, model=currModel, type="MagAzim",
+	domain="longLatGrid", output="SpatRaster", polecrop=FALSE))
 expect_equal(terra::ext(llgrastOrig), terra::ext(c(-180.5, 180.5, -90.5,90.5)))
 expect_equal(dim(llgrastOrig)[3], 3)
 expect_equal(names(llgrastOrig), c("magnitude", "azimuth", "plateid"))
 
 ## east_north
 ## data.frame
-expect_silent(llg2 <- velocities(age=30, model=currModel, type="east_north", domain="longLatGrid"))
+expect_silent(llg2 <- velocities("static_polygons", age=30, model=currModel, type="east_north", domain="longLatGrid"))
 
 expect_true(inherits(llg2, "data.frame"))
 expect_equal(ncol(llg2),5)
@@ -73,21 +79,21 @@ expect_equal(llg[c(1,2,5)], llg2[c(1,2,5)])
 
 
 ## Raster - output
-expect_silent(llgrastENdef <- velocities(age=30, model=currModel, type="east_north",
+expect_silent(llgrastENdef <- velocities("static_polygons", age=30, model=currModel, type="east_north",
 	domain="longLatGrid", output="SpatRaster"))
 expect_true(inherits(llgrastENdef, "SpatRaster"))
 
-### resample=TRUE
-expect_silent(llgrastEN <- velocities(age=30, model=currModel, type="east_north",
-	domain="longLatGrid", output="SpatRaster", cellraster=TRUE))
+### polecrop=TRUE
+expect_silent(llgrastEN <- velocities("static_polygons", age=30, model=currModel, type="east_north",
+	domain="longLatGrid", output="SpatRaster", polecrop=TRUE))
 expect_equal(llgrastEN, llgrastENdef)
-expect_equal(dim(llgrastEN)[3], 2)
-expect_equal(names(llgrastEN), c("east", "north"))
-expect_equal(terra::ext(llgrast), terra::ext(c(-180, 180, -90,90)))
+expect_equal(dim(llgrastEN)[3], 3)
+expect_equal(names(llgrastEN), c("east", "north", "plateid"))
+expect_equal(terra::ext(llgrast), terra::ext(c(-180.5, 180.5, -89.5,89.5)))
 
-### resample=FALSE
-expect_silent(llgrastENorig <- velocities(age=30, model=currModel, type="east_north",
-	domain="longLatGrid", output="SpatRaster", cellraster=FALSE))
+### polecrop=FALSE
+expect_silent(llgrastENorig <- velocities("static_polygons", age=30, model=currModel, type="east_north",
+	domain="longLatGrid", output="SpatRaster", polecrop=FALSE))
 expect_equal(dim(llgrastENorig)[3], 3)
 expect_equal(names(llgrastENorig), c("east", "north", "plateid"))
 expect_equal(terra::ext(llgrastENorig), terra::ext(c(-180.5, 180.5, -90.5,90.5)))
@@ -95,14 +101,14 @@ expect_equal(terra::ext(llgrastENorig), terra::ext(c(-180.5, 180.5, -90.5,90.5))
 # II. healpix-domain (data.frame output)
 
 ## Raster - output should provide error
-expect_error(velocities(age=30, model=currModel, type="east_north",
+expect_error(velocities("static_polygons", age=30, model=currModel, type="east_north",
 	domain="healpix", output="SpatRaster"))
 
-expect_error(velocities(age=30, model=currModel, type="MagAzim",
+expect_error(velocities("static_polygons", age=30, model=currModel, type="MagAzim",
 	domain="healpix", output="SpatRaster"))
 
 ## MagAzim
-expect_silent(hpMA<- velocities(age=30, model=currModel, type="MagAzim",
+expect_silent(hpMA<- velocities("static_polygons", age=30, model=currModel, type="MagAzim",
 	domain="healpix"))
 expect_true(inherits(hpMA, "data.frame"))
 expect_equal(ncol(hpMA),5)
@@ -119,7 +125,7 @@ expect_true(min(hpMA$magnitude)>=0)
 expect_true(max(hpMA$magnitude)<=200)
 
 # east-north
-expect_silent(hpEN<- velocities(age=30, model=currModel, type="east_north",
+expect_silent(hpEN<- velocities("static_polygons", age=30, model=currModel, type="east_north",
 	domain="healpix"))
 expect_true(inherits(hpEN, "data.frame"))
 expect_equal(ncol(hpEN),5)
@@ -137,16 +143,15 @@ expect_equal(hpMA[c(1,2,5)], hpEN[c(1,2,5)])
 ########################################----------------------------------------
 # Misc
 # Incorrect domain
-expect_error(velocities(age=30, model=currModel, type="MagAzim",
+expect_error(velocities("static_polygons", age=30, model=currModel, type="MagAzim",
 	domain="garbage", output="data.frame"))
 
 # Incorrect type
-expect_error(velocities(age=30, model=currModel, type="garbage",
+expect_error(velocities("static_polygons", age=30, model=currModel, type="garbage",
 	domain="longLatGrid", output="data.frame"))
 
 # Target age beyond the limit suggested by gws
-velocities(age=30, model="SETON2012", type="MagAzim")
-# Iterate for ages
+expect_error(velocities("static_polygons", age=230, model="SETON2012", type="MagAzim"))
 
 
 ################################################################################